feature/GH-2420: move SiteContextParamService to services folder
