+++ projects/core/src/site-context/providers/site-context-params-providers.ts
@@ -1,5 +1,5 @@
 import { APP_INITIALIZER, Provider } from '@angular/core';
-import { SiteContextParamsService } from '../facade/site-context-params.service';
+import { SiteContextParamsService } from '../services/site-context-params.service';
 import { UrlSerializer } from '@angular/router';
 import { SiteContextUrlSerializer } from '../services/site-context-url-serializer';
 import { SiteContextRoutesHandler } from '../services/site-context-routes-handler';
