+++ projects/core/src/site-context/services/site-context-url-serializer.ts
@@ -1,6 +1,6 @@
 import { DefaultUrlSerializer, UrlTree } from '@angular/router';
 import { Injectable } from '@angular/core';
-import { SiteContextParamsService } from '../facade/site-context-params.service';
+import { SiteContextParamsService } from './site-context-params.service';
 import { SiteContextConfig } from '../config/site-context-config';
 
 export interface ParamValuesMap {
